move maverick out into test directory
